requirements.txt
schedule~=1.2.1
python-dotenv~=1.0.1
pika~=1.3.2
requests~=2.31.0